requirements.txt
pandas
google-generativeai
openpyxl